+++ packages/headless/doc-parser/doc-parser.ts
@@ -12,6 +12,7 @@ import {
   Engine,
   resolveEngine,
 } from './src/headless-export-resolvers/engine-resolver';
+import {productRecommendationUseCase} from './use-cases/product-recommendation';
 import {recommendationUseCase} from './use-cases/recommendation';
 import {searchUseCase} from './use-cases/search';
 import {UseCaseConfiguration} from './use-cases/use-case-configuration';
@@ -25,7 +26,7 @@ interface UseCase {
 interface ResolvedUseCase {
   name: string;
   controllers: Controller[];
-  actionLoaders: ActionLoader[];
+  actions: ActionLoader[];
   engine: Engine;
 }
 
@@ -40,6 +41,11 @@ const useCases: UseCase[] = [
     entryFile: 'temp/recommendation.api.json',
     config: recommendationUseCase,
   },
+  {
+    name: 'product-recommendation',
+    entryFile: 'temp/product-recommendation.api.json',
+    config: productRecommendationUseCase,
+  },
 ];
 
 function resolveUseCase(useCase: UseCase): ResolvedUseCase {
@@ -51,13 +57,13 @@ function resolveUseCase(useCase: UseCase): ResolvedUseCase {
   const controllers = config.controllers.map((controller) =>
     resolveController(entryPoint, controller)
   );
-  const actionLoaders = config.actionLoaders.map((loader) =>
+  const actions = config.actionLoaders.map((loader) =>
     resolveActionLoader(entryPoint, loader)
   );
 
   const engine = resolveEngine(entryPoint, config.engine);
 
-  return {name, controllers, actionLoaders, engine};
+  return {name, controllers, actions, engine};
 }
 
 const resolved = useCases.map(resolveUseCase);