+++ packages/quantic/force-app/main/default/lwc/quanticHeadlessLoader/__tests__/quanticHeadlessLoader.test.js
@@ -1,30 +1,36 @@
 /* eslint-disable jest/expect-expect */
 import {
-  setEngineConfiguration,
+  setInitializedCallback,
+  setEngineOptions,
   registerComponentForInit,
   setComponentInitialized,
-  getHeadlessEngine,
+  getHeadlessEnginePromise,
   initializeWithHeadless
 } from '../quanticHeadlessLoader';
 import { CoveoHeadlessStub } from '../../testUtils/coveoHeadlessStub';
-import {Deferred} from '../../quanticUtils/quanticUtils';
+import { Deferred } from '../../quanticUtils/quanticUtils';
 
 describe('c/quanticHeadlessLoader', () => {
-  const testConfig = {
-    organizationId: 'testOrganization',
-    accessToken: 'bogus-token-xxxxx-xxxxx',
-    search: {
-      pipeline: 'default',
-      searchHub: 'default',
+  const testOptions = {
+    configuration: {
+      organizationId: 'testOrganization',
+      accessToken: 'bogus-token-xxxxx-xxxxx',
+      search: {
+        pipeline: 'default',
+        searchHub: 'default',
+      }
     }
   };
 
   let resolvedTestConfig;
+  let mockedConsoleWarn;
   let mockedConsoleError;
   let initialize;
   let testId = 'search-page-1';
 
   let testElement;
+
+  let mockEngineConstructor;
 
   const createComponentEntryWithInitStatus = (isInitialized) => ({
     element: document.createElement('div'),
@@ -40,13 +46,16 @@ describe('c/quanticHeadlessLoader', () => {
 
   beforeEach(() => {
     jest.useFakeTimers();
-    mockedConsoleError = jest.fn();
     initialize = jest.fn();
+    mockEngineConstructor = jest.fn().mockReturnValue(CoveoHeadlessStub.buildSearchEngine());
+    mockedConsoleWarn = jest.fn();
+    mockedConsoleError = jest.fn();
+    console.warn = mockedConsoleWarn;
     console.error = mockedConsoleError;
     global.CoveoHeadless = CoveoHeadlessStub;
     testElement = document.createElement('div');
     resolvedTestConfig = new Deferred();
-    resolvedTestConfig.resolve(testConfig);
+    resolvedTestConfig.resolve(testOptions);
   });
 
   afterEach(() => {
@@ -55,12 +64,27 @@ describe('c/quanticHeadlessLoader', () => {
     jest.restoreAllMocks();
   });
 
-  describe('setEngineConfiguration', () => {
-    describe('when coveoHeadless is undefined', () => {
-      it('should define coveoHeadless and set config', async () => {
-        setEngineConfiguration(testConfig, testId, testElement);
+  describe('setInitializedCallback', () => {
+    beforeEach(() => {
+      window.coveoHeadless = {
+        [testId]: {}
+      };
+    });
 
-        expect(await window.coveoHeadless[testId].config.promise).toBe(testConfig);
+    it('should define callback for specified engine', async () => {
+      const testCallback = () => { };
+      setInitializedCallback(testCallback, testId);
+
+      expect(await window.coveoHeadless[testId].initializedCallback).toBe(testCallback);
+    });
+  });
+
+  describe('setEngineOptions', () => {
+    describe('when coveoHeadless is undefined', () => {
+      it('should define coveoHeadless and set options', async () => {
+        setEngineOptions(testOptions, mockEngineConstructor, testId, testElement);
+
+        expect(await window.coveoHeadless[testId].options.promise).toBe(testOptions);
       });
     });
 
@@ -69,32 +93,30 @@ describe('c/quanticHeadlessLoader', () => {
         window.coveoHeadless = {};
       });
 
-      it('should set config', async () => {
-        setEngineConfiguration(testConfig, testId, testElement);
+      it('should set options', async () => {
+        setEngineOptions(testOptions, mockEngineConstructor, testId, testElement);
 
-        expect(await window.coveoHeadless[testId].config.promise).toBe(testConfig);
+        expect(await window.coveoHeadless[testId].options.promise).toBe(testOptions);
       });
 
-      describe('when a config is already preset', () => {
+      describe('when options are already preset', () => {
         beforeEach(() => {
           window.coveoHeadless[testId] = {
-            config: resolvedTestConfig
+            options: resolvedTestConfig
           }
         });
 
-        it('should thow an error', () => {
-          const newConfig = {
-            organizationId: 'otherOrganization',
-            accessToken: 'bogus-token-xxxxx-ooooo',
-          }  
-          let caughtError;
-          try {
-            setEngineConfiguration(newConfig, testId, testElement);
-          } catch(error) {
-            caughtError = error.message;
+        it('should log a warning', () => {
+          const newOptions = {
+            configuration: {
+              organizationId: 'otherOrganization',
+              accessToken: 'bogus-token-xxxxx-ooooo',
+            }
           }
 
-          expect(caughtError).toContain(`Attempted to overwrite configuration for engine: ${testId}`);
+          setEngineOptions(newOptions, mockEngineConstructor, testId, testElement);
+
+          expect(mockedConsoleWarn).toBeCalledWith(`Attempted to overwrite engine options for engine ID: ${testId}`)
         });
       })
     });
@@ -129,20 +151,20 @@ describe('c/quanticHeadlessLoader', () => {
       describe('when no search is pending', () => {
         it('should register the component', () => {
           registerComponentForInit(testElement, testId);
-  
+
           assertComponentIsRegistered(testElement);
         });
-  
+
         it('should not register the component if it already is', () => {
           window.coveoHeadless[testId].components.push({
-              element: testElement,
-              initialized: false
+            element: testElement,
+            initialized: false
           });
           const mockedPush = jest.fn();
           window.coveoHeadless[testId].components.push = mockedPush;
-  
+
           registerComponentForInit(testElement, testId);
-  
+
           expect(mockedPush).not.toBeCalled();
         });
       });
@@ -156,13 +178,13 @@ describe('c/quanticHeadlessLoader', () => {
       });
     });
 
-    describe('when coveoHeadless is defined', () => {      
+    describe('when coveoHeadless is defined', () => {
       beforeEach(() => {
         window.coveoHeadless = {
           [testId]: {
             components: [],
             engine: Promise.resolve(CoveoHeadlessStub.buildSearchEngine()),
-            config: new Deferred()
+            options: new Deferred()
           }
         }
       });
@@ -180,8 +202,13 @@ describe('c/quanticHeadlessLoader', () => {
 
         it('should set the component to initialized', async () => {
           setComponentInitialized(testElement, testId);
-  
+
           assertComponentIsSetInitialized(testElement, testId);
+        });
+
+        it('should not execute initialization callback', async () => {
+          setComponentInitialized(testElement, testId);
+
           jest.runAllTimers();
           const engine = await window.coveoHeadless[testId].engine;
           expect(engine.executeFirstSearch).not.toBeCalled();
@@ -199,55 +226,70 @@ describe('c/quanticHeadlessLoader', () => {
           ];
         });
 
-        it('should set the component to initialized and start delayed search', async () => {
+        it('should set the component to initialized', async () => {
           setComponentInitialized(testElement, testId);
-  
+
           assertComponentIsSetInitialized(testElement, testId);
-          jest.runAllTimers();
-          await window.coveoHeadless.engine;
-          const engine = await window.coveoHeadless[testId].engine;
-          expect(engine.executeFirstSearch).toBeCalled();
+        });
+
+        describe('when initialized callback is set', () => {
+          const callbackMock = jest.fn();
+          beforeEach(() => {
+            window.coveoHeadless[testId].initializedCallback = callbackMock;
+          })
+
+          it('should set the component to initialized and execute delayed initalized callback', async () => {
+            setComponentInitialized(testElement, testId);
+
+            assertComponentIsSetInitialized(testElement, testId);
+            jest.runAllTimers();
+            await window.coveoHeadless.engine;
+            expect(callbackMock).toBeCalled();
+          });
         });
       });
     });
   });
 
-  describe('getHeadlessEngine', () => {
-    describe('when the engine is undefined with a set config', () => {      
+  describe('getHeadlessEnginePromise', () => {
+    describe('when the engine is undefined with a set options', () => {
       beforeEach(() => {
         window.coveoHeadless = {
           [testId]: {
             components: [],
-            config: resolvedTestConfig,
+            options: resolvedTestConfig,
+            engineConstructor: mockEngineConstructor,
+            bindings: {}
           }
         }
       });
 
-      it('should init the engine return a promise that resolves to that instance', async () => {
-        const engine = await getHeadlessEngine(testId);
-
+      it('should init the engine and return a promise', async () => {
+        const engine = await getHeadlessEnginePromise(testId);
+        expect(mockEngineConstructor).toBeCalled();
         expect(engine).toBeTruthy();
       });
     });
 
-    describe('when the engine is undefined without a set config', () => {      
+    describe('when the engine is undefined without a set options', () => {
       beforeEach(() => {
         window.coveoHeadless = {
           [testId]: {
             components: [],
+            bindings: {}
           }
         }
       });
 
-      it('should init the engine return a promise that resolves to that instance', async () => {
+      it('should throw an error specifying options have not been set', async () => {
         let caughtError;
         try {
-          await getHeadlessEngine(testId);
-        } catch(error) {
+          await getHeadlessEnginePromise(testId);
+        } catch (error) {
           caughtError = error.message;
         }
 
-        expect(caughtError).toContain('Engine configuration has not been set.');
+        expect(caughtError).toContain('Engine options have not been set.');
       });
     });
 
@@ -260,13 +302,15 @@ describe('c/quanticHeadlessLoader', () => {
         window.coveoHeadless = {
           [testId]: {
             components: [],
-            engine: Promise.resolve(definedEngine)
+            enginePromise: Promise.resolve(definedEngine),
+            engineConstructor: mockEngineConstructor,
+            bindings: {},
           }
         }
       });
 
       it('should return a promise that resolves to an instance of the engine', async () => {
-        const engine = await getHeadlessEngine(testId);
+        const engine = await getHeadlessEnginePromise(testId);
 
         expect(engine).toEqual(definedEngine);
       });
@@ -274,7 +318,7 @@ describe('c/quanticHeadlessLoader', () => {
   });
 
   describe('initializeWithHeadless', () => {
-    describe('when the engine is undefined', () => {      
+    describe('when the engine is undefined', () => {
       beforeEach(() => {
         window.coveoHeadless = {
           [testId]: {
@@ -282,14 +326,16 @@ describe('c/quanticHeadlessLoader', () => {
               element: testElement,
               initialized: false,
             }],
-            config: resolvedTestConfig,
+            engineConstructor: mockEngineConstructor,
+            options: resolvedTestConfig,
+            bindings: {},
           }
         }
       });
 
       it('should init the engine and initialize ', async () => {
         await initializeWithHeadless(testElement, testId, initialize);
-        const engine = await window.coveoHeadless[testId].engine;
+        const engine = await window.coveoHeadless[testId].enginePromise;
 
         expect(engine).toBeTruthy();
         expect(initialize).toHaveBeenCalled();
@@ -299,16 +345,16 @@ describe('c/quanticHeadlessLoader', () => {
       describe('when initializing the engine fails', () => {
         const errorMessage = 'simulating failure';
         beforeEach(() => {
-          window.coveoHeadless[testId].config = new Deferred();
-          window.coveoHeadless[testId].config.reject(errorMessage);
+          window.coveoHeadless[testId].options = new Deferred();
+          window.coveoHeadless[testId].options.reject(errorMessage);
         });
-        
-        it ('should throw an error', async () => {
+
+        it('should throw an error', async () => {
           let caughtError;
           initializeWithHeadless(testElement, testId, initialize);
           try {
-            await window.coveoHeadless[testId].engine;
-          } catch(error) {
+            await window.coveoHeadless[testId].enginePromise;
+          } catch (error) {
             caughtError = error.message;
           }
 
@@ -328,15 +374,17 @@ describe('c/quanticHeadlessLoader', () => {
               element: testElement,
               initialized: false
             }],
-            config: resolvedTestConfig,
-            engine: Promise.resolve(definedEngine),
+            options: resolvedTestConfig,
+            enginePromise: Promise.resolve(definedEngine),
+            engineConstructor: mockEngineConstructor,
+            bindings: {},
           }
         }
       });
 
       it('should initialize the component using the defined engine', async () => {
         await initializeWithHeadless(testElement, testId, initialize);
-        const engine = await window.coveoHeadless[testId].engine;
+        const engine = await window.coveoHeadless[testId].enginePromise;
 
         expect(engine).toEqual(definedEngine);
         expect(initialize).toHaveBeenCalled();